(front) add document signatures page
