+++ frontend/src/app/document-editor/document-content/anexo/anexo.component.html
@@ -1,0 +1,39 @@
+<div class="pb-4">
+    <div class="header d-flex justify-content-between">
+        <b class="header-title">Anexo {{form.get('index').value!}}</b> 
+        <button class="ghost-btn" nbButton ghost nbTooltip="Eliminar anexo" (click)="deleteAnexo()">
+            <nb-icon icon="close-outline" status="danger"></nb-icon>
+        </button>
+    </div>
+    <hr>
+    <ng-container [formGroup]="form">
+        <div class="form-group">
+            <!--label class="label col-form-label">Título</label>
+            <input type="text" fullWidth nbInput formControlName="title" placeholder="Ingresar título">
+            <label class="label col-form-label">Subtítulo</label>
+            <input type="text" fullWidth nbInput formControlName="subtitle" placeholder="Ingresar subtítulo"-->
+            
+            <!--label class="label col-form-label">Fuente del contenido</label>
+            <nb-select fullWidth placeholder="Seleccionar fuente"  (selectedChange)="contentSourceOnChange()" [selected]="anexoContentSource">
+                <nb-option value="attachedFile" (click)="anexoContentSource = 'attachedFile'">Pdf</nb-option>
+                <nb-option value="manual" (click)="anexoContentSource = 'manual'">Edición manual</nb-option>
+            </nb-select-->
+        </div>
+        <!--div class="d-flex align-items-center" *ngIf="anexoContentSource == 'attachedFile'"-->
+        <div class="d-flex align-items-center">
+            <!--div class="form-group"-->    
+            <input *ngIf="!file && !uploading" type="file" class="file-input" (change)="onFileSelected($event)" #fileUpload nbInput fullWidth>
+            <ng-container *ngIf="file && !uploading">
+                Archivo subido: {{file.filename}}
+                <button class="ghost-btn ml-2" nbButton ghost nbTooltip="Eliminar archivo" (click)="deleteFile()">
+                    <nb-icon icon="close-outline" status="danger"></nb-icon>
+                </button>
+            </ng-container>
+            <nb-card class="ghost-card" *ngIf="uploading" [nbSpinner]="true" nbSpinnerSize="small" nbSpinnerStatus="primary">
+                <nb-card-body></nb-card-body>
+            </nb-card>
+            <!--/div-->
+        </div>
+        <!--app-text-editor *ngIf="anexoContentSource == 'manual'" formControlName="content"></app-text-editor-->
+    </ng-container>
+</div>
