+++ frontend/src/app/document-editor/document-content/anexo/anexo.component.scss
@@ -1,0 +1,13 @@
+.ghost-card {
+    margin: 0;
+    border: 0;
+}
+
+.ghost-btn {
+    height: fit-content;
+    width: fit-content;
+}
+
+.header-title {
+    padding-top: 12px;
+}
