+++ frontend/src/app/document-editor/document-content/document-content.component.scss
@@ -1,0 +1,62 @@
+.label {
+    padding-left: 0;
+}
+
+.full-width {
+    width: 100%;
+}
+
+.ghost-btn {
+    height: fit-content;
+    width: fit-content;
+}
+
+.ghost-card {
+    margin: 0;
+    border: 0;
+}
+
+.error-dialog {
+    max-width: 40rem;
+}
+
+.error-dialog nb-card-footer {
+    text-align: right;
+}
+
+.action-result {
+    color: var(--text-primary-color);
+    font-weight: 550;
+}
+
+.fade-out {
+    animation: fadeOut ease 2.5s;
+    animation-fill-mode: forwards;
+    animation-delay: 4s;
+}
+
+@keyframes fadeOut {
+    0% {
+        opacity: 1;
+    }
+    
+    100% {
+        opacity: 0;
+    }
+}
+
+.flex-container {
+    flex: 1;
+    display: flex;
+    flex-flow: column;
+    height: calc(100vh - 12rem);
+}
+
+.scrollable {
+    overflow-y: auto;
+    flex-grow: 1; 
+}
+
+.scrollable nb-card:last-child {
+    margin-bottom: 0;
+}
