+++ frontend/src/app/document-editor/document-content/init-settings-dialog/init-settings-dialog.component.html
@@ -1,0 +1,26 @@
+<nb-card class="h-100" *ngIf="viewState == 'loading'" [nbSpinner]="true" nbSpinnerStatus="primary">
+    <nb-card-body></nb-card-body>
+  </nb-card>
+  <nb-card *ngIf="viewState == 'rendering'">
+    <nb-card-header>Crear documento</nb-card-header>
+    <nb-card-body>
+        <div class="form-group">  
+            <label class="label col-form-label">Dependencia emisora</label>
+            <nb-select fullWidth placeholder="Seleccionar emisor" [(selected)]="selectedIssuerId">
+                <nb-option *ngFor="let item of issuers" [value]="item.id">{{item.description}}</nb-option>
+            </nb-select>
+        </div>
+        <div class="form-group">
+            <label class="label col-form-label">Tipo de documento</label>
+            <nb-select fullWidth placeholder="Seleccionar tipo" [(selected)]="selectedDocumentTypeId">
+            <nb-option *ngFor="let item of documentTypes" [value]="item.id">{{item.description}}</nb-option>
+            </nb-select>
+        </div>
+    </nb-card-body>
+    <nb-card-footer>
+    <button class="mr-2" nbButton status="primary" (click)="close(true)">Aceptar</button>
+    <button nbButton (click)="close(false)">Cerrar</button>
+    </nb-card-footer>
+</nb-card>
+
+
