+++ frontend/src/app/document-editor/document-signatures/signature-selector/signature-selector.component.html
@@ -1,0 +1,15 @@
+<nb-card class="h-100" *ngIf="viewState == 'loading'" [nbSpinner]="true" nbSpinnerStatus="primary">
+    <nb-card-body></nb-card-body>
+  </nb-card>
+<nb-card class="h-100" *ngIf="viewState == 'rendering'">
+    <nb-card-header>Agregar firma/sello</nb-card-header>
+    <nb-card-body>
+        <nb-select fullWidth placeholder="Seleccionar sello/firma" [formControl]="stampIdFormControl">
+            <nb-option *ngFor="let stamp of stamps" [value]="stamp.id">{{stamp.description}}</nb-option>
+        </nb-select>
+    </nb-card-body>
+    <nb-card-footer>
+        <button nbButton status="primary" (click)="submit()" class="mr-4">Guardar</button>
+        <button nbButton (click)="cancel()">Cancelar</button>
+    </nb-card-footer>
+</nb-card>
