+++ frontend/src/app/document-editor/document-signatures/signature-selector/signature-selector.component.scss
@@ -1,0 +1,4 @@
+.content {
+    height: calc(100vh - 14rem);
+}
+
